Renamed keymap component to keymap-edit
