+++ src/components/keymap/index.ts
@@ -1,4 +1,4 @@
-export * from './keymap.component';
 export * from './keymap.routes';
 export * from './add/keymap-add.component';
+export * from './edit/keymap-edit.component';
 export * from './header/keymap-header.component';
